# src/main/resources/config-openapi/openapi.yaml
openapi: 3.0.3
info:
  title: User Microservice API
  description: Microservicio para gestión de usuarios y roles
  version: 1.0.0
servers:
  - url: http://localhost:8084/api/v1
    description: User Microservice
paths:
...
  /user-roles:
    get:
      tags:
        - UserRole
      summary: Obtener todas las relaciones usuario-rol
      operationId: getAllUserRoles
      responses:
        '200':
          description: Relaciones obtenidas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserRoleResponseDTO'
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'

    post:
      tags:
        - UserRole
      summary: Asignar un rol a un usuario
      operationId: assignRoleToUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRoleRequestDTO'
      responses:
        '201':
          description: Rol asignado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRoleResponseDTO'
        '400':
          description: Solicitud inválida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /user-roles/{id}:
    get:
      tags:
        - UserRole
      summary: Obtener relación usuario-rol por ID
      operationId: getUserRoleById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Relación encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRoleResponseDTO'
        '404':
          description: Relación no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'

    delete:
      tags:
        - UserRole
      summary: Eliminar relación usuario-rol
      operationId: deleteUserRole
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Relación eliminada exitosamente
        '404':
          description: Relación no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /user-roles/user/{userId}:
    get:
      tags:
        - UserRole
      summary: Obtener todos los roles asignados a un usuario
      operationId: getRolesByUserId
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Roles del usuario
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserRoleResponseDTO'
        '404':
          description: Usuario no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /user-roles/role/{roleId}:
    get:
      tags:
        - UserRole
      summary: Obtener todos los usuarios asignados a un rol
      operationId: getUsersByRoleId
      parameters:
        - name: roleId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Usuarios con el rol especificado
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserRoleResponseDTO'
        '404':
          description: Rol no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string

    # DTOs de Role
...
    UserRoleRequestDTO:
      type: object
      required:
        - userId
        - roleId
      properties:
        userId:
          type: integer
          format: int64
          description: ID del usuario asociado al rol.
        roleId:
          type: integer
          format: int64
          description: ID del rol asignado al usuario.
    UserRoleResponseDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: ID del registro de relación usuario-rol.
        userId:
          type: integer
          format: int64
          description: ID del usuario asociado.
        roleId:
          type: integer
          format: int64
          description: ID del rol asociado.
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        active:
          type: boolean